Modify the representation of string lengths in the EE storage layer so that 1 byte is used for strings 63 characters or less and 4-bytes are used to store the length of all other strings. NValue now caches the length of the string in bytes 8-11 and the length of the length preceding value in byte 12. The null terminator at the end of strings is also eliminated. The wire format for arrays in parameter sets is also changed so that byte array parameters use a 4-byte length instead of a 2-byte length. Other array types continue to use a 2-byte length. This is to allow blobs/strings to passed as params to stored procedures without the expensive Java String construction that can entail several copies. The maximum object length is set to 1 megabyte and the maximum row length is set to 2 megabytes. This should allow a single max size blob to be retrieved in a row along with associated metadata. Also add a BlobTorture benchmark client that loads blobs to each partition and then selects them randomly. This is for ENG-435.

+++ src/ee/common/ValuePeeker.hpp
@@ -75,7 +75,7 @@ public:
         return value.getObjectValue();
     }
 
-    static inline int16_t peekObjectLength(const NValue value) {
+    static inline int32_t peekObjectLength(const NValue value) {
         assert(value.getValueType() == VALUE_TYPE_VARCHAR);
         return value.getObjectLength();
     }
