+++ src/ee/plannodes/receivenode.h
@@ -68,9 +68,9 @@ class ReceivePlanNode : public AbstractPlanNode
     std::vector<ValueType>& getOutputColumnTypes();
     const std::vector<ValueType>& getOutputColumnTypes() const;
 
-    void setOutputColumnSizes(std::vector<uint16_t>& sizes);
-    std::vector<uint16_t>& getOutputColumnSizes();
-    const std::vector<uint16_t>& getOutputColumnSizes() const;
+    void setOutputColumnSizes(std::vector<int32_t>& sizes);
+    std::vector<int32_t>& getOutputColumnSizes();
+    const std::vector<int32_t>& getOutputColumnSizes() const;
 
     std::string debugInfo(const std::string& spacer) const;
 
@@ -92,7 +92,7 @@ protected:
     std::vector<int> m_outputColumnGuids;
     std::vector<std::string> m_outputColumnNames;
     std::vector<ValueType> m_outputColumnTypes;
-    std::vector<uint16_t> m_outputColumnSizes;
+    std::vector<int32_t> m_outputColumnSizes;
 };
 
 }
